refactor: move RomTester
